requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-MySQL db==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
Werkzeug==3.1.3
scapy==2.5.0